+++ CHANGELOG.md
@@ -10,6 +10,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### VB -> C#
 
+
+### C# -> VB
+
+
+## [8.3.1] - 2021-08-22
+
+
+### Vsix
+
+
+### VB -> C#
+
 * Convert `orderby distinct` in linq [#736](https://github.com/icsharpcode/CodeConverter/issues/736)
 * Convert nested Select queries in linq [#635](https://github.com/icsharpcode/CodeConverter/issues/635)
 * `Chr` converted to `Strings.Chr` where code page aware conversion needed [#745](https://github.com/icsharpcode/CodeConverter/issues/745)